+++ NEWS.md
@@ -11,7 +11,7 @@
 - ResultParser was not correctly called in EventStudyAPI
 - ResultParser was not able to download files from url
 - getTaskStatus did not work properly in EventStudyAPI 
-- fix av and avyc paring
+- fix av and avyc parsing
 
 
 ## Version 0.30